pyproject.toml
[project]
name = "platform"
version = "0.1.0"
description = ""
authors = [
    { name = "Your Name", email = "you@example.com" }
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard] (>=0.117.1,<0.118.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "sqlmodel (>=0.0.25,<0.0.26)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "minio (>=7.2.17,<8.0.0)",
    "redis (>=6.4.0,<7.0.0)",
    "sqlalchemy-datatables (>=2.0.1,<3.0.0)",
    "perfetto (>=0.14.0,<0.15.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "psycopg[binary] (>=3.2.10,<4.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
packages = [{ include = "src" }]